iommu: msm: Correct the data type mis-matches in VA
MSM IOMMU driver presently supports mapping of virtual
addresses of 32-bit long only. Because of this, 32-bit
long virtual address was okay and some mis-matches were
silent. This is harmless but still buggy. Also, for
64-bit virtual address mapping, some API needs update.

Make input virtual address always 'unsigned long' and
trunk (if necessary) based on the page table format.

+++ drivers/iommu/msm_iommu_pagetable.c
@@ -374,7 +374,7 @@ static inline int is_fully_aligned(unsigned int va, phys_addr_t pa, size_t len,
 static int __msm_iommu_pagetable_map_range(struct msm_iommu_pt *pt,
 		       unsigned long va, void *cookie,
 		       struct msm_iommu_map_ops *ops,
-		       unsigned int len, int prot)
+		       size_t len, int prot)
 {
 	phys_addr_t pa;
 	unsigned int start_va = va;
@@ -600,7 +600,7 @@ void msm_iommu_pagetable_unmap_range(struct msm_iommu_pt *pt, unsigned long va,
 }
 
 int msm_iommu_pagetable_map_range(struct msm_iommu_pt *pt, unsigned long va,
-		struct scatterlist *sg, unsigned int len, int prot)
+		struct scatterlist *sg, size_t len, int prot)
 {
 	return __msm_iommu_pagetable_map_range(pt, va, sg, &sg_ops, len, prot);
 }
